move index.js into new-xostor-form.js
